# Backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for a project.",
        "title": "GoTransact",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api",
    "paths": {
...
        "/protected/post-payment": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new payment request with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Create a new payment request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization header with bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Payment Request Input",
                        "name": "paymentInput",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/utils.PostPaymentInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully created payment request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": true
                },
                "messsage": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "utils.PostPaymentInput": {
            "type": "object",
            "required": [
                "amount",
                "cardnumber",
                "cvv",
                "expirydate"
            ],
            "properties": {
                "amount": {
                    "type": "string"
                },
                "cardnumber": {
                    "type": "string"
                },
                "cvv": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "expirydate": {
                    "type": "string"
                }
            }
        },
...
    }
}